Feat : 타입 수정, 오타 제거

+++ backend/hr/src/main/java/com/bbap/hr/dto/CustomUserDetails.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class CustomUserDetails {
-    private Long empId;
+    private Integer empId;
 }
