+++ backend/hr/src/main/java/com/bbap/hr/dto/WorkplaceDto.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Setter
 @Builder
 public class WorkplaceDto {
-    private Long workplaceId;
+    private Integer workplaceId;
     private String workplaceName;
 }
