+++ backend/hr/src/main/java/com/bbap/hr/entity/WorkplaceEntity.java
@@ -22,7 +22,7 @@ public class WorkplaceEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="workplace_id")
-    private Long workplaceId;
+    private Integer workplaceId;
 
     @Column(name="workplace_name")
     private String workplaceName;
